1. Default build type to Debug - since this is still a work-in-progress 2. Fixed and now working - transfer size implementation, where chunk is transferred over in a smaller portions, to avoid high memory usage 3. Moved from DEBUG to MY_DEBUG 4. initParameters should've been long since it's long in the main.cpp

+++ file.h
@@ -17,7 +17,7 @@ class File {
     unsigned long chunkSize_;
     unsigned long transferSize_;
   public:
-    File(std::string path, int InitParameters, unsigned long threads);
+    File(std::string path, long InitParameters, unsigned long threads);
 
     void setPath(std::string path);
     std::string getPath() const;
